+++ apps/backend/src/oauth/providers/microsoft.tsx
@@ -12,11 +12,12 @@ export class MicrosoftProvider extends OAuthBaseProvider {
   static async create(options: {
     clientId: string,
     clientSecret: string,
+    microsoftTenantId?: string,
   }) {
     return new MicrosoftProvider(...await OAuthBaseProvider.createConstructorArgs({
-      issuer: "https://login.microsoftonline.com",
-      authorizationEndpoint: "https://login.microsoftonline.com/consumers/oauth2/v2.0/authorize",
-      tokenEndpoint: "https://login.microsoftonline.com/consumers/oauth2/v2.0/token",
+      issuer: `https://login.microsoftonline.com${"/" + options.microsoftTenantId || ""}`,
+      authorizationEndpoint: `https://login.microsoftonline.com/${options.microsoftTenantId || 'consumers'}/oauth2/v2.0/authorize`,
+      tokenEndpoint: `https://login.microsoftonline.com/${options.microsoftTenantId || 'consumers'}/oauth2/v2.0/token`,
       redirectUri: getEnvVariable("STACK_BASE_URL") + "/api/v1/auth/oauth/callback/microsoft",
       baseScope: "User.Read",
       ...options,