+++ examples/partial-prerendering/CHANGELOG.md
@@ -1,5 +1,12 @@
 # partial-prerendering
 
+## 2.5.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.30
+
 ## 2.5.29
 
 ### Patch Changes