+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack
 
+## 2.5.30
+
+### Patch Changes
+
+- More OAuth providers
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.30
+  - @stackframe/stack-ui@2.5.30
+  - @stackframe/stack-sc@2.5.30
+
 ## 2.5.29
 
 ### Patch Changes